fix: correctly tell R to install from BioConductor
Apparently, the way to specify that you have some BioConductor dependency is to add an empty `biocViews:` node to DESCRIPTION (https://stackoverflow.com/questions/62486902/how-to-automatically-install-bioconductor-package-on-which-a-cran-package-depend).

I'd rather specify the dependency in `Remotes:` so that only the specified packages are installed from BioConductor.
But I wasn't able to figure out how to do that.

Not quite sure why the previous version worked locally.