fix: use sub-second precision
Rounding down to the nearest second was done implicitly by fuzzyjoin (at least that was what led to identical results).

Also: avoid segments ending up in two intervals. With sub-second precision, the probability that a segment would start at exactly the boundary of an intervals is extremely low but still. The new condition segment_start in [start, end), not in [start, end]