make add_lena_stats take intervals in the same format as other functions
For some reason, I decided that this one function should have `start` and `end` columns?
Probably I wanted the column boundary columns to be named the same whether we were using wall or wav time.
But the wall time is not implemented yet and I'd rather make my function deal with the column names than the user.