test, temp: in make_five_min_approximation test non-empty intervals only
As we are moving from the population of intervals that have segments in them to the one of all possible intervals within the recordings, we will get more intervals as some of the new ones won't have any segments in them.
We don't want the test to fail for that reason, so we will temporarily switch to only testing intervals where at least one metric is not zero: we don't really care about the intervals without any measurable events in them - even if they contain segments.