0.19.0: fix public version identification
Function `is_public_version` in `get_data.R` would correctly identify versions starting with "0" or ending with "dev" as not public but it then identified anything else as public because I had "v1.0.0" hard-coded in the check which, of course, satisfied the criteria for being public.
Basically, the rule was: if it isn't formatted as a dev version then it's public which works until I introduce a v0.0.0.9000 tag somewhere by mistake and it is identified as public.