implement getting subsets
Use `subset` parameter for all random, random+top-5 VI+TD-VI, or all annotations.

extra: add expect_column_contents test helper function. This is not a full testthat custom expectation but it'll do. I am still not sure if it is the best idea to have test-helpers.R in the package code, I did it because that is what Hadley recommended on some GitHub (?) thread. I don't like that I have code that uses functions that aren't blabr's dependencies (testthat and digest). Specifically for custom expectation functions, there is a recommended file to put them in inside the testthat folder. See testthat's docs on custom expectations for more information.