feat: update fixations_to_timeseries
- Drop keepCols. It was used in every single call I could find and thus can be easily replaced by a `select` in the future.
- `bin_size` -> `t_step`. This is selfish: I find it much easier to think about this procedure as resampling a piecewise constant function. Honestly, the "bins" were fine, it is "binning" and "binify" that I found misleading because it only makes sense during downsampling,
- + docs.
- + readability.
- Moved expandFixList inside the function and renamed it to expand_single_fixation.

Extra: The last step was done by left-joining timepoints to the (possibly trimmed) fixations on fixation_id? This was confusing because timepoints can (very rarely) have fewer fixations than the (possibly trimmed) fixations but it is impossible for the opposite to occur. Changing that `left_join` to an `inner_join` thus didn't actually change anything.