feat: speed up fixations_to_timeseries
Instead of converting each fixation to a dataframe of time points one by one, we are now joining the fixations list with the list of all potential timepoints which is an order of magnitude faster.

This function can be further sped up by not doing `select` at the end of the first join and getting rid of the second join altogether.