fix: bug in deduplicating time points
This bug is very unlikely to lead to any changes in practice. This is just me being extra neurotic.

subset(
    timepoints,
    time_bin[2:length(time_bin)] != time_bin[1:(length(time_bin)-1)])

The mask (second argument of `subset`) is 1 element shorter than `timepoints` and is thus recycled - the first element in the mask is used to decide whether both the first and the last element of `timepoints` stay. We never want to remove the last row from the dataframe because if there are duplicates, we always drop the first one - and the last row can never be the first of the duplicates. 99.99% percent of the time we aren't going to remove it because the very first timepoint is not a duplicate. However, if it was, we would lose the very last time point. Again, this is very unlikely to be the case so I don't expect this change to affect any of our datasets.